fix(ainic): add copyright headers and fix data type mismatches
AMD SMI returns uint64_t for NIC statistics but the AI NIC sampling
code was using uint32_t, which could cause data truncation for
high-throughput network counters.

Changes:
- Add SPDX copyright headers to AI NIC source files
- Change NIC stats fields from uint32_t to uint64_t to match AMD SMI API
- Remove redundant amdsmi.h include (already included via header)

+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/ainic_stats.cpp
@@ -1,3 +1,6 @@
+// Copyright (c) Advanced Micro Devices, Inc.
+// SPDX-License-Identifier: MIT
+
 #include "ainic_stats.hpp"
 
 #include <memory>
@@ -182,20 +185,20 @@ ai_nic_stats_collector::update_data_for_one_nic(amdsmi_processor_handle processo
                                                 &num_stats, stats.get());
 
             const std::unordered_map<std::string_view,
-                                     std::function<void(nic_stats&, uint32_t)>>
+                                     std::function<void(nic_stats&, uint64_t)>>
                 stat_handlers = {
                     { nic_stats::RX_RDMA_UCAST_BYTES,
-                      [](nic_stats& d, uint32_t v) { d._rx_rdma_ucast_bytes = v; } },
+                      [](nic_stats& d, uint64_t v) { d._rx_rdma_ucast_bytes = v; } },
                     { nic_stats::RX_RDMA_UCAST_PKTS,
-                      [](nic_stats& d, uint32_t v) { d._rx_rdma_ucast_pkts = v; } },
+                      [](nic_stats& d, uint64_t v) { d._rx_rdma_ucast_pkts = v; } },
                     { nic_stats::TX_RDMA_UCAST_BYTES,
-                      [](nic_stats& d, uint32_t v) { d._tx_rdma_ucast_bytes = v; } },
+                      [](nic_stats& d, uint64_t v) { d._tx_rdma_ucast_bytes = v; } },
                     { nic_stats::TX_RDMA_UCAST_PKTS,
-                      [](nic_stats& d, uint32_t v) { d._tx_rdma_ucast_pkts = v; } },
+                      [](nic_stats& d, uint64_t v) { d._tx_rdma_ucast_pkts = v; } },
                     { nic_stats::RX_RDMA_CNP_PKTS,
-                      [](nic_stats& d, uint32_t v) { d._rx_rdma_cnp_pkts = v; } },
+                      [](nic_stats& d, uint64_t v) { d._rx_rdma_cnp_pkts = v; } },
                     { nic_stats::TX_RDMA_CNP_PKTS,
-                      [](nic_stats& d, uint32_t v) { d._tx_rdma_cnp_pkts = v; } },
+                      [](nic_stats& d, uint64_t v) { d._tx_rdma_cnp_pkts = v; } },
                 };
 
             // Retrieve relevant stats.
