+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/amd_smi_ainic.hpp
@@ -1,3 +1,6 @@
+// Copyright (c) Advanced Micro Devices, Inc.
+// SPDX-License-Identifier: MIT
+
 #pragma once
 
 #include "core/common.hpp"
@@ -61,12 +64,12 @@ struct nic_data
 private:
     std::string _nic;
     uint32_t    _nic_index        = 0;
-    uint32_t    _rx_rdma_cnp_pkts = 0;
-    uint32_t    _tx_rdma_cnp_pkts = 0;
-    uint32_t    _rx_ucast_bytes   = 0;
-    uint32_t    _tx_ucast_bytes   = 0;
-    uint32_t    _rx_ucast_pkts    = 0;
-    uint32_t    _tx_ucast_pkts    = 0;
+    uint64_t    _rx_rdma_cnp_pkts = 0;
+    uint64_t    _tx_rdma_cnp_pkts = 0;
+    uint64_t    _rx_ucast_bytes   = 0;
+    uint64_t    _tx_ucast_bytes   = 0;
+    uint64_t    _rx_ucast_pkts    = 0;
+    uint64_t    _tx_ucast_pkts    = 0;
 };
 
 }  // namespace amd_smi
